refactor: consolidate parse-utils into normalize, remove dead preview_url, rename server_tts_proxy
1. Delete dashboard/src/api/parse-utils.ts — move asStringList to
   normalize.ts, update 2 importers (board.ts, dashboard.ts) to import
   from normalize.ts directly. Eliminates semantic divergence between
   parse-utils.asString (no trim, always returns string) and
   normalize.asString (trims, returns undefined).

2. Remove dead preview_url field from voice_config.ml — the
   /api/v1/voice/tts route has no handler (trpg_tts_proxy was removed
   as dead code). Advertising a 404 URL misleads clients.

3. Rename server_tts_proxy.ml to server_voice_config.ml — the module
   only contains voice_config_payload after TTS proxy removal.

Closes #3122.
